+++ src/main/java/engine/model/physicalEngine/movement/Velocity.java
@@ -1,27 +1,27 @@
 package engine.model.physicalEngine.movement;
 
 public class Velocity {
-    private float velocityX;
-    private float velocityY;
+    private double velocityX;
+    private double velocityY;
 
-    public Velocity(float velocityX, float velocityY) {
+    public Velocity(double velocityX, double velocityY) {
         this.velocityX = velocityX;
         this.velocityY = velocityY;
     }
 
-    public float getVelocityX() {
+    public double getVelocityX() {
         return velocityX;
     }
 
-    public float getVelocityY() {
+    public double getVelocityY() {
         return velocityY;
     }
 
-    public void setVelocityX(float velocityX) {
+    public void setVelocityX(double velocityX) {
         this.velocityX = velocityX;
     }
 
-    public void setVelocityY(float velocityY) {
+    public void setVelocityY(double velocityY) {
         this.velocityY = velocityY;
     }
 
@@ -30,7 +30,7 @@ public class Velocity {
         setVelocityY(velocityY);
     }
 
-    public void addVelocity(float bonusX, float bonusY) {
+    public void addVelocity(double bonusX, double bonusY) {
         setVelocityX(getVelocityX() + bonusX);
         setVelocityY(getVelocityY() + bonusY);
     }
